Cell widget and keyboard widget
The cell widget will carry the value of each cell in the board.

The user will use the Keyboard widget to select the value of a cell.

When the user clicks a cell, the cell is attached to the keyboard. So, every action on the keyboard will trigger changes on the attached cell.